+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change Log
 
 ## [Unreleased]
+
+## [v0.3] – 2017-02-14
 ### Added
 - Create `generateCuboid` for 3D rectangle.
+- Create `generateSphere`.
 
 ## [v0.2] – 2017-02-14
 ### Changed
@@ -18,6 +21,7 @@
 ### Added
 - Create functions `generateCylinder`, `generateLine`, `generateRectangle`.
 
-[Unreleased]: https://github.com/joshforisha/fractal-noise-js/compare/v0.2...HEAD
+[Unreleased]: https://github.com/joshforisha/fractal-noise-js/compare/v0.3...HEAD
+[v0.3]: https://github.com/joshforisha/fractal-noise-js/compare/v0.2...v0.3
 [v0.2]: https://github.com/joshforisha/fractal-noise-js/compare/v0.1...v0.2
 [v0.1]: https://github.com/joshforisha/fractal-noise-js/releases/tag/v0.1